Baja de Vehiculos
Falta chequear si el auto esta asignado a un chofer.

+++ TPCuatrimestal-Equipo13/Negocio/DomicilioNegocio.cs
@@ -46,7 +46,7 @@ namespace Negocio
             }
         }
 
-        public void BajaDomicilio(int IdDomicilio)
+        public void BajaDomicilio(long IdDomicilio)
         {
             AccesoDatos datos = new AccesoDatos();
